+++ website/src/displaySecret/DisplaySecret.tsx
@@ -2,8 +2,8 @@ import * as React from 'react';
 import { useEffect, useState, useCallback } from 'react';
 import { useLocation, useParams } from 'react-router-dom';
 import Error from './Error';
-import Form from './Form';
-import { decryptMessage } from './utils';
+import Form from '../createSecret/Form';
+import { decryptMessage } from '../utils/utils';
 import { useTranslation } from 'react-i18next';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faCopy } from '@fortawesome/free-solid-svg-icons';
