+++ website/src/displaySecret/Error.tsx
